Refactored stair number code
